+++ TribulaCustomServerParent/TribulaCustomServer-Spigot/src/main/java/net/minecraft/server/BlockPressurePlateWeighted.java
@@ -1,6 +1,6 @@
 package net.minecraft.server;
 
-import org.bukkit.event.entity.EntityInteractEvent; // CraftBukkit
+import org.bukkit.event.entity.EntityInteractEvent;
 
 public class BlockPressurePlateWeighted extends BlockPressurePlateAbstract {
 
@@ -13,7 +13,7 @@ public class BlockPressurePlateWeighted extends BlockPressurePlateAbstract {
 
     protected BlockPressurePlateWeighted(Material material, int i, MaterialMapColor materialmapcolor) {
         super(material, materialmapcolor);
-        this.w(this.blockStateList.getBlockData().set(BlockPressurePlateWeighted.POWER, Integer.valueOf(0)));
+        this.w(this.blockStateList.getBlockData().set(BlockPressurePlateWeighted.POWER, 0));
         this.weight = i;
     }
 
@@ -62,11 +62,11 @@ public class BlockPressurePlateWeighted extends BlockPressurePlateAbstract {
     }
 
     protected int getPower(IBlockData iblockdata) {
-        return iblockdata.get(BlockPressurePlateWeighted.POWER).intValue();
+        return iblockdata.get(BlockPressurePlateWeighted.POWER);
     }
 
     protected IBlockData a(IBlockData iblockdata, int i) {
-        return iblockdata.set(BlockPressurePlateWeighted.POWER, Integer.valueOf(i));
+        return iblockdata.set(BlockPressurePlateWeighted.POWER, i);
     }
 
     public int a(World world) {
@@ -74,11 +74,11 @@ public class BlockPressurePlateWeighted extends BlockPressurePlateAbstract {
     }
 
     public IBlockData fromLegacyData(int i) {
-        return this.getBlockData().set(BlockPressurePlateWeighted.POWER, Integer.valueOf(i));
+        return this.getBlockData().set(BlockPressurePlateWeighted.POWER, i);
     }
 
     public int toLegacyData(IBlockData iblockdata) {
-        return iblockdata.get(BlockPressurePlateWeighted.POWER).intValue();
+        return iblockdata.get(BlockPressurePlateWeighted.POWER);
     }
 
     protected BlockStateList getStateList() {
